Fix infinite recursion when calling setSeed with long

+++ math/src/main/java/smile/math/random/MersenneTwister.java
@@ -75,7 +75,7 @@ public class MersenneTwister implements RandomNumberGenerator {
 
     @Override
     public void setSeed(long seed) {
-        setSeed(seed % UniversalGenerator.BIG_PRIME);
+        setSeed((int)(seed % UniversalGenerator.BIG_PRIME));
     }
 
     public void setSeed(int seed) {
